moved concurrency helpers to daemon
